Moved start service calls to session from connection as each session has its own services.
