+++ sdl_android_lib/src/com/smartdevicelink/SdlConnection/SdlSession.java
@@ -1,25 +1,39 @@
 package com.smartdevicelink.SdlConnection;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.io.PipedInputStream;
+import java.io.PipedOutputStream;
 import java.util.concurrent.CopyOnWriteArrayList;
 
+import android.annotation.SuppressLint;
+import android.os.Build;
 import android.util.Log;
+import android.view.Surface;
 
+import com.smartdevicelink.encoder.SdlEncoder;
 import com.smartdevicelink.exception.SdlException;
 import com.smartdevicelink.protocol.ProtocolMessage;
 import com.smartdevicelink.protocol.enums.ServiceType;
 import com.smartdevicelink.protocol.heartbeat.IHeartbeatMonitor;
 import com.smartdevicelink.protocol.heartbeat.IHeartbeatMonitorListener;
 import com.smartdevicelink.proxy.LockScreenManager;
+import com.smartdevicelink.proxy.RPCRequest;
+import com.smartdevicelink.streaming.AbstractPacketizer;
+import com.smartdevicelink.streaming.IStreamListener;
+import com.smartdevicelink.streaming.StreamPacketizer;
+import com.smartdevicelink.streaming.StreamRPCPacketizer;
 import com.smartdevicelink.transport.BaseTransportConfig;
 import com.smartdevicelink.transport.enums.TransportType;
 
-public class SdlSession implements ISdlConnectionListener, IHeartbeatMonitorListener {
+public class SdlSession implements ISdlConnectionListener, IHeartbeatMonitorListener, IStreamListener {
 	private static CopyOnWriteArrayList<SdlConnection> shareConnections = new CopyOnWriteArrayList<SdlConnection>();
 	
 	SdlConnection _sdlConnection = null;
 	private byte sessionId;
-	@SuppressWarnings("unused")
-    private byte wiproProcolVer;
+    @SuppressWarnings("unused")
+	private byte wiproProcolVer;
 	private ISdlConnectionListener sessionListener;
 	private BaseTransportConfig transportConfig;
     IHeartbeatMonitor _outgoingHeartbeatMonitor = null;
@@ -27,7 +41,14 @@ public class SdlSession implements ISdlConnectionListener, IHeartbeatMonitorList
     private static final String TAG = "SdlSession";
     private LockScreenManager lockScreenMan  = new LockScreenManager();
 
+	AbstractPacketizer mPacketizer = null;
     
+	StreamRPCPacketizer mRPCPacketizer = null;
+	StreamPacketizer mVideoPacketizer = null;
+	StreamPacketizer mAudioPacketizer = null;
+	SdlEncoder mSdlEncoder = null;    
+    
+	private final static int BUFF_READ_SIZE = 1024;
 	
 	public static SdlSession createSession(byte wiproVersion, ISdlConnectionListener listener, BaseTransportConfig btConfig) {
 		
@@ -104,6 +125,18 @@ public class SdlSession implements ISdlConnectionListener, IHeartbeatMonitorList
 		return "";
 	}
 	
+	public void startService (ServiceType serviceType, byte sessionID) {
+		if (_sdlConnection == null) 
+			return;
+		
+		_sdlConnection.startService(serviceType, sessionID);			
+	}
+	
+	public void endService (ServiceType serviceType, byte sessionID) {
+		if (_sdlConnection == null) 
+			return;
+		_sdlConnection.endService(serviceType, sessionID);	
+	}
 	
 	public void startSession() throws SdlException {
 		SdlConnection connection = null;
@@ -148,6 +181,203 @@ public class SdlSession implements ISdlConnectionListener, IHeartbeatMonitorList
 			return false;
 		return _sdlConnection != null && _sdlConnection.getIsConnected();
 	}
+	
+	public Surface createOpenGLInputSurface(int frameRate, int iFrameInterval, int width,
+			int height, int bitrate, ServiceType sType, byte rpcSessionID) {
+		try {
+			PipedOutputStream stream = (PipedOutputStream) startStream(sType, rpcSessionID);
+			if (stream == null) return null;
+			mSdlEncoder = new SdlEncoder();
+			mSdlEncoder.setFrameRate(frameRate);
+			mSdlEncoder.setFrameInterval(iFrameInterval);
+			mSdlEncoder.setFrameWidth(width);
+			mSdlEncoder.setFrameHeight(height);
+			mSdlEncoder.setBitrate(bitrate);
+			mSdlEncoder.setOutputStream(stream);
+		} catch (IOException e) {
+			return null;
+		}
+		return mSdlEncoder.prepareEncoder();
+	}
+	
+	public void startEncoder () {
+		if(mSdlEncoder != null) {
+		   mSdlEncoder.startEncoder();
+		}
+	}
+
+	public void releaseEncoder() {
+		if(mSdlEncoder != null) {
+		   mSdlEncoder.releaseEncoder();
+		}
+	}
+	
+	public void drainEncoder(boolean endOfStream) {
+		if(mSdlEncoder != null) {
+		   mSdlEncoder.drainEncoder(endOfStream);
+		}
+	}
+
+	@Override
+	public void sendStreamPacket(ProtocolMessage pm) {
+		if (_sdlConnection == null) 
+			return;
+		sendMessage(pm);
+	}	
+	
+	
+	public void startStream(InputStream is, ServiceType sType, byte rpcSessionID) throws IOException {
+        if (sType.equals(ServiceType.NAV))
+        {
+        	mVideoPacketizer = new StreamPacketizer(this, is, sType, rpcSessionID, this);
+        	mVideoPacketizer.sdlConnection = this.getSdlConnection();
+        	mVideoPacketizer.start();
+        }
+        else if (sType.equals(ServiceType.PCM))
+        {
+        	mAudioPacketizer = new StreamPacketizer(this, is, sType, rpcSessionID, this);
+        	mAudioPacketizer.sdlConnection = this.getSdlConnection();
+        	mAudioPacketizer.start();            	
+        }
+	}
+	
+	@SuppressLint("NewApi") public OutputStream startStream(ServiceType sType, byte rpcSessionID) throws IOException {
+			OutputStream os = new PipedOutputStream();
+			InputStream is = null;
+			if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.GINGERBREAD) {
+				is = new PipedInputStream((PipedOutputStream) os, BUFF_READ_SIZE);
+			} else {
+				is = new PipedInputStream((PipedOutputStream) os);
+			}	        
+	        if (sType.equals(ServiceType.NAV))
+	        {
+	            mVideoPacketizer = new StreamPacketizer(this, is, sType, rpcSessionID, this);
+	            mVideoPacketizer.sdlConnection = this.getSdlConnection();
+	            mVideoPacketizer.start();
+	        }       
+	        else if (sType.equals(ServiceType.PCM))
+	        {
+	        	mAudioPacketizer = new StreamPacketizer(this, is, sType, rpcSessionID, this);
+	        	mAudioPacketizer.sdlConnection = this.getSdlConnection();
+	        	mAudioPacketizer.start();            	
+	        }
+	        else
+	        {
+	        	os.close();
+	        	is.close();
+	        	return null;
+	        }            
+						
+			return os;
+	}
+		
+	public void startRPCStream(InputStream is, RPCRequest request, ServiceType sType, byte rpcSessionID, byte wiproVersion) {
+		try {
+	        mPacketizer = new StreamRPCPacketizer(null, this, is, request, sType, rpcSessionID, wiproVersion, 0, this);
+			mPacketizer.start();
+		} catch (Exception e) {
+	        Log.e("SdlConnection", "Unable to start streaming:" + e.toString());
+	    }
+	}
+	
+	public OutputStream startRPCStream(RPCRequest request, ServiceType sType, byte rpcSessionID, byte wiproVersion) {
+		try {
+			OutputStream os = new PipedOutputStream();
+	        InputStream is = new PipedInputStream((PipedOutputStream) os);
+			mPacketizer = new StreamRPCPacketizer(null, this, is, request, sType, rpcSessionID, wiproVersion, 0, this);
+			mPacketizer.start();
+			return os;
+		} catch (Exception e) {
+	        Log.e("SdlConnection", "Unable to start streaming:" + e.toString());
+	    }
+		return null;
+	}
+	
+	public void pauseRPCStream()
+	{
+		if (mPacketizer != null)
+		{
+			mPacketizer.pause();
+		}
+	}
+	
+	public void resumeRPCStream()
+	{
+		if (mPacketizer != null)
+		{
+			mPacketizer.resume();
+		}
+	}
+	
+	public void stopRPCStream()
+	{
+		if (mPacketizer != null)
+		{
+			mPacketizer.stop();
+		}
+	}
+	
+	public boolean stopAudioStream()
+	{
+		if (mAudioPacketizer != null)
+		{
+			mAudioPacketizer.stop();
+			return true;
+		}
+		return false;
+	}
+	
+	public boolean stopVideoStream()
+	{
+		if (mVideoPacketizer != null)
+		{
+			mVideoPacketizer.stop();
+			return true;
+		}
+		return false;
+	}
+	
+	public boolean pauseAudioStream()
+	{
+		if (mAudioPacketizer != null)
+		{
+			mAudioPacketizer.pause();
+			return true;
+		}
+		return false;
+	}
+	
+	public boolean pauseVideoStream()
+	{
+		if (mVideoPacketizer != null)
+		{
+			mVideoPacketizer.pause();
+			return true;
+		}
+		return false;
+	}
+	
+	public boolean resumeAudioStream()
+	{
+		if (mAudioPacketizer != null)
+		{
+			mAudioPacketizer.resume();
+			return true;
+		}
+		return false;		
+	}
+	
+	public boolean resumeVideoStream()
+	{
+		if (mVideoPacketizer != null)
+		{
+			mVideoPacketizer.resume();
+			return true;
+		}
+		return false;
+	}	
+		
+		
 
 	@Override
 	public void onTransportDisconnected(String info) {
