+++ sdl_android_lib/src/com/smartdevicelink/streaming/AbstractPacketizer.java
@@ -3,6 +3,7 @@ package com.smartdevicelink.streaming;
 import java.io.IOException;
 import java.io.InputStream;
 
+import com.smartdevicelink.SdlConnection.SdlSession;
 import com.smartdevicelink.protocol.enums.ServiceType;
 import com.smartdevicelink.proxy.RPCRequest;
 
@@ -12,9 +13,10 @@ abstract public class AbstractPacketizer {
 	private final static int BUFF_READ_SIZE = 1000000;
 	protected byte _rpcSessionID = 0;
 	
-	protected ServiceType _session = null;
 	protected InputStream is = null;
 	protected byte[] buffer = new byte[BUFF_READ_SIZE];
+	protected ServiceType _serviceType = null;
+	protected SdlSession _session = null;
 	protected boolean upts = false;
 	protected RPCRequest _request = null;
 	protected byte _wiproVersion = 1;
@@ -22,20 +24,22 @@ abstract public class AbstractPacketizer {
 	//protected long ts = 0, intervalBetweenReports = 5000, delta = 0;
 	protected long intervalBetweenReports = 5000, delta = 0;
 
-	public AbstractPacketizer(IStreamListener streamListener, InputStream is, ServiceType sType, byte rpcSessionID) throws IOException {
+	public AbstractPacketizer(IStreamListener streamListener, InputStream is, ServiceType sType, byte rpcSessionID, SdlSession session) throws IOException {
         this._streamListener = streamListener;
 		this.is = is;
 		_rpcSessionID = rpcSessionID;
-		_session = sType;
+		_serviceType = sType;
+		this._session = session;
 	}
 	
-	public AbstractPacketizer(IStreamListener streamListener, InputStream is, RPCRequest request, ServiceType sType, byte rpcSessionID, byte wiproVersion) throws IOException {
+	public AbstractPacketizer(IStreamListener streamListener, InputStream is, RPCRequest request, ServiceType sType, byte rpcSessionID, byte wiproVersion, SdlSession session) throws IOException {
         this._streamListener = streamListener;
 		this.is = is;
 		_rpcSessionID = rpcSessionID;
-		_session = sType;
+		_serviceType = sType;
 		_request = request;
 		_wiproVersion = wiproVersion;
+		this._session = session;
 	}	
 
 	public abstract void start() throws IOException;
